Refactor: ui 패키지 추가, 콘솔 입출력 관련 클래스 이동
