+++ src/main/java/racingcar/ui/ConsoleInputReader.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.ui;
 
 import java.util.Scanner;
