+++ src/main/java/racingcar/ui/ConsoleOutputMessage.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.ui;
 
 public enum ConsoleOutputMessage {
