+++ src/main/java/racingcar/ui/ConsoleOutputWriter.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.ui;
 
 public class ConsoleOutputWriter {
